LayoutTests/js/dfg-to-primitive-pass-symbol-expected.txt

Tests that op_to_primitive pass Symbol.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


str concat generates op_to_primitive.
PASS doToPrimitive() threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
